requirements.txt
# Core
streamlit==1.53.1
groq==1.0.0
python-dotenv==1.0.0

# Whisper (transcription locale)
openai-whisper==20250625
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0

# TTS (doublage)
qwen-tts>=0.1.0
soundfile>=0.12.0
numpy>=1.24.0
tqdm>=4.65.0

# Optionnel: Flash Attention pour GPU (accélération)
# pip install flash-attn --no-build-isolation
